demos/03-agent-framework/01-basics/agentfw_basics-py/pyproject.toml
[project]
name = "agentfw-basics-py"
version = "0.1.0"
description = "Agent Framework basics demo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "azure-identity>=1.16.0",
    "agent-framework-azure-ai",
    "colorama>=0.4.6",
]
